Fix: Deleted/Moved tasks re-appeared on sync
RemoteTasks were not handled correctly when their local
task had been unmapped to them. In this case, all RemoteTasks
ended up colliding due to the use of HashMaps and hardcoded
invalid id of "-99".

RemoteTaskLists are not given a pseudo-id on delete, and hence do
not suffer from this issue.

Fixes #272

Signed-off-by: Jonas Kalderstam <jonas@kalderstam.se>